3d_dicom_software_for_students_streamlit/requirements.txt
streamlit>=1.38.0
dicom2stl>=0.5.0
numpy-stl>=3.1.1
plotly>=5.24.0
openai>=1.51.0
python-dotenv>=1.0.1
streamlit-plotly-events>=0.0.6
